add gl interactivity, refactor code
