+++ README.md
@@ -5,4 +5,7 @@ Start server on http://localhost:3000
 npm run dev
 ```
 
-code from https://webgl2fundamentals.org/webgl/lessons/webgl-fundamentals.html
+## Resources
+WebGL code: https://webgl2fundamentals.org/webgl/lessons/webgl-fundamentals.html
+Kaleidoscope shader example (Unity): https://danielilett.com/2020-02-19-tut3-8-crazy-kaleidoscopes/
+OpenGL: https://learnopengl.com/
