+++ src/main.ts
@@ -1,27 +1,21 @@
-import { resizeCanvasToDisplaySize, createShader, createProgram } from './util';
-
-// allowed by the flags provided to esbuild, ignore err
+// ignore compiler errors
+// allowed by the flags provided to esbuild
 import vertexShaderSource from './shaders/shader.vert'; // --loader:.vert=text flag 
 import fragmentShaderSource from './shaders/shader.frag'; // --loader:.frag=text flag 
 
+import * as util from './util';
+import * as webglUtil from './webgl_utils'
+
 function main() {
-  const canvas = document.getElementById("canvas") as HTMLCanvasElement | null;
-  if (canvas == null) throw new Error("Couldn't find canvas element.");
+  const canvas = document.querySelector("#canvas") as HTMLCanvasElement | null;
+  if (!canvas) throw new Error("Could not find canvas element in document");
 
-  const gl = canvas.getContext("webgl2");
-  if (gl == null) throw new Error("Couldn't get webgl2 context from canvas.");
-
-  const vertexShader = createShader(gl, gl.VERTEX_SHADER, vertexShaderSource);
-  const fragmentShader = createShader(gl, gl.FRAGMENT_SHADER, fragmentShaderSource);
-
-  if (vertexShader == undefined) throw new Error("Vertex shader is undefined.");
-  if (fragmentShader == undefined) throw new Error("Fragment shader is undefined.");
-
-  var program = createProgram(gl, vertexShader, fragmentShader);
-
-  if (program == undefined) throw new Error("WebGL2 program is undefined.");
+  const gl = webglUtil.getWebGL2Context(canvas);
+  const program = webglUtil.createWebGLProgramFromSource(gl, [vertexShaderSource, fragmentShaderSource]);
 
   const positionAttributeLocation = gl.getAttribLocation(program, "a_position");
+  const colorAUniformLocation = gl.getUniformLocation(program, "color_a");
+
   const positionBuffer = gl.createBuffer();
   gl.bindBuffer(gl.ARRAY_BUFFER, positionBuffer);
 
@@ -45,19 +39,40 @@ function main() {
     positionAttributeLocation, size, type, normalize, stride, attrib_offset
   )
 
-  resizeCanvasToDisplaySize(gl.canvas as HTMLCanvasElement | null)
+  var color_a = 0.5;
 
-  gl.viewport(0, 0, gl.canvas.width, gl.canvas.height);
-  gl.clearColor(0, 0, 0, 0);
-  gl.clear(gl.COLOR_BUFFER_BIT);
+  document.addEventListener("keydown", (event) => {
+    switch (event.key) {
+      case "w":
+        color_a = Math.min(color_a + 0.1, 1.0);
+        break;
+      case "s":
+        color_a = Math.max(color_a - 0.1, 0.0);
+        break;
+    }
+    drawScene();
+  })
 
-  gl.useProgram(program);
-  gl.bindVertexArray(vao)
+  drawScene();
 
-  const primitiveType = gl.TRIANGLES;
-  const offset = 0;
-  const count = 3;
-  gl.drawArrays(primitiveType, offset, count);
+  function drawScene() {
+    util.resizeCanvasToDisplaySize(gl.canvas as HTMLCanvasElement | null)
+
+    gl.viewport(0, 0, gl.canvas.width, gl.canvas.height);
+
+    gl.clearColor(0, 0, 0, 0);
+    gl.clear(gl.COLOR_BUFFER_BIT | gl.DEPTH_BUFFER_BIT);
+
+    gl.useProgram(program);
+    gl.bindVertexArray(vao)
+
+    gl.uniform1f(colorAUniformLocation, color_a);
+
+    const primitiveType = gl.TRIANGLES;
+    const offset = 0;
+    const count = 3;
+    gl.drawArrays(primitiveType, offset, count);
+  }
 }
 
 main();
