+++ style.css
@@ -6,10 +6,11 @@ body {
 
 div {
   display: block;
-  background-color: aliceblue;
 }
 
 canvas {
-  background-color: aqua;
   align-self: center;
+
+  width: 100%;
+  height: 100%;
 }
